tests for the put end point
